requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
fastapi==0.63.0
fastapi-mqtt==0.3.0
gmqtt==0.6.9
h11==0.12.0
idna==2.10
paho-mqtt==1.5.1
pydantic==1.8.1
redis==3.5.3
requests==2.25.1
starlette==0.13.6
typing-extensions==3.10.0.0
urllib3==1.26.4
uvicorn==0.13.4
